refactor(model): consolidate Measurement-related code
