--- internal/engine/model/model.go
@@ -1,7 +1,0 @@
-// Package model defines shared data structures and interfaces.
-package model
-
-const (
-	// DefaultProbeIP is the default probe IP.
-	DefaultProbeIP = "127.0.0.1"
-)
